tagging release 0.3
=Changes=
Added a file (doc_macros.xml)  containing  target macros. Replaced
the contents of doc.xml with a generic installation file which needs
to be edited when creating a new manual (see INSTALL).              
Java XML validation (Validate class) was replaced  with ant XML     
validation. For this  ant-nodeps.jar was added in the lib folder and the
Validate class was removed from doc-utils.jar.  The validation now shows
the file with the error (line, column) instead of a line and column in  
tmp/main.xml. Also, the build stops if the XML file is invalid.         


Fixes for
http://bugs.activeeon.com/browse/DOC-33
http://bugs.activeeon.com/browse/DOC-9




git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/documentation/tags/docchain_0.3@11159 28e8926c-6b08-0410-baaa-805c5e19b8d6