Added the possibility to insert the xml header (<?xml ... ?>) when we are extracting an XML or a fractal file. It was not possible before since we cannot put the start annotation tag before this header.
From now on, using the start annotation @snippet-start-with-header <Name of your snippet>, snippet extractor will add
the header before extracting. If you do not want to add it, just do as usual using the
@snippet-start <Name of your snippet> annotation.

NB: end, break and resume annotations do not have to be changed. If you do so, you will have an exception from
the snippet extractor saying that your file is not valid.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/documentation/trunk@13667 28e8926c-6b08-0410-baaa-805c5e19b8d6