Added the possibility to extract code snippets from several directories.
Instead of giving a single path as the first argument of the Snippetizer.main() method, from now on, we can give several paths separated by ':' (like a unix classpath).
The second argument which is the target directory is unchanged.

Renamed the previous snippets.start property as snippets.start.default. This property will be only used if the snippets.start property is not set
and contains the previous path that is, ${basedir}/../src.

Removed the pa.xml file which is not used anymore.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/documentation/trunk@14462 28e8926c-6b08-0410-baaa-805c5e19b8d6