Fixed DOC-122: tutorial-start annotations should not be located before copyright (for copyright update issues) and tutorial extraction should add it automatically
Using @tutorial-start-with-copyright or @snippet-start-with-copyright, it will automatically insert the copyright before the tutorial file or the snippet extraction.

Added a way to automatically add the copyright to snippet and tutorial files.
This is to turn round the fact that we cannot put an annotation before copyrights anymore (copyright updating issue).

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/documentation/trunk@14830 28e8926c-6b08-0410-baaa-805c5e19b8d6