Fixed DOC-128: toolchain/doc_path.xml has been edited into the scheduling projects. It should be rewritten into the documentation project and updated after into each project.
This file has not been simply copied from the sheduling project to the documentation one as some of the modifaction were dependent on the sheduler project.
The main cause of this modification was the setting of properties depending on which target we've called (Scheduler vs rm).
As ant properties are not mutable and as variables are discouraged, these dependinding properties are now set only at the target call time through a macro
instead of being set at the import time.

Thus, the main modifications are the following:
- Removed these properties from the doc_paths.xml path
- Added a macro in doc_marcos.xml to set these properties
- Called this macro inside each doc.xml target that needs the setting of these properties

This macro assumes that the doc.name property has been set before calling it.
The docs.src.dir and docs.mainFile properties should also be set in the project side.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/documentation/trunk@14889 28e8926c-6b08-0410-baaa-805c5e19b8d6